refactor(chain-store-zainodb): rename ports to adapter and split it up
The module was called `ports` but defined none. The driven ports live in the
`zaino-chain-store` domain crate; what was here is ten `impl <port> for
DbReader<T>`/`FinalisedState<T>` blocks and the conversions they need. An
adapter has no ports of its own — it satisfies someone else's — so the name
read as "the port contracts are defined here", which inverts the layering the
crate exists to express, and sent anyone looking for a contract to the wrong
file.

It had also reached 1927 lines across eight section dividers: error mapping,
scalar conversions, block conversions, capability and schema mapping, the port
impls, the write-direction conversions, and a feature-gated port. One file, one
merge-conflict surface, and sections related only by having nowhere else to go.

Now `adapter.rs` — the port impls and the read instrumentation — over:

- `adapter/error_map.rs`, this backend's failures as the domain names them
- `adapter/to_domain.rs`, reading: what is on disk as the domain names it
- `adapter/from_domain.rs`, writing: what the domain hands over as this stores it
- `adapter/history.rs`, the feature-gated transparent-history port, so its gate
  is one `mod` declaration rather than a `cfg` on each method

The two conversion directions are not named `into_domain`/`into_stored` as
suggested, because `StoredBlock` is a *domain* type: a module called
`into_stored` would hold the conversions *out of* this backend, the opposite of
what it says. `to_domain` and `from_domain` name the direction and cannot be
read backwards.

Tests moved with what they test rather than staying in one module, so the
round-trip pairs sit beside their conversions. All 18 are preserved and now
run as `adapter::*`.

Visibility was re-derived rather than widened: items used only within their new
module stayed private, and `pub(super)` was added only where the compiler
required it. The three items consumers actually import — `domain_outpoint`,
`stored_tx_out`, `indexed_block_from_stored` — are re-exported from `adapter`,
so callers change only the module name.

Also fixes a call site the typo rename missed
(`tests/finalised_state/migrations/v1_1_to_v1_2.rs:291`), which left
`--all-targets` failing to build.
