+++ packages/zaino-chain-store-zainodb/src/store.rs
@@ -247,7 +247,7 @@ use zebra_chain::parameters::NetworkKind;
 #[cfg(feature = "prometheus")]
 use crate::metric_names::*;
 
-use crate::ports::domain_block_ref;
+use crate::adapter::domain_block_ref;
 use crate::store::{finalised_source::v1::DB_VERSION_V1, router::EphemeralMode};
 use crate::types::{BlockHash, ChainWork, Height, IndexedBlock, GENESIS_HEIGHT};
 use zaino_chain_store::ChainStoreConfig;
